refactoring Jobs and JobController
Make Worker not part of the Houses, instead worker are part of the worker controller and do all kinds of jobs.

Not finished yet. Still refactoring the Worker Script to get Jobs from the Job Controller and refactoring the JobController so that you can get different Jobs from it. (Drinking Eating, Wood collenting)

Aiming for Resources as Objects on map.